+++ src/batched/KokkosBatched_Util.hpp
@@ -295,60 +295,6 @@ struct Mode {
   };
 };
 
-#if !defined(KOKKOS_IF_ON_HOST)
-
-template <class>
-struct algo_level3_blocked_mb_impl;
-template <>
-struct algo_level3_blocked_mb_impl<Kokkos::HostSpace> {
-  static constexpr int value = 4;
-};
-#if defined(KOKKOS_ENABLE_CUDA)
-template <>
-struct algo_level3_blocked_mb_impl<Kokkos::CudaSpace> {
-  static constexpr int value = 2;
-};
-#endif
-#if defined(KOKKOS_ENABLE_HIP)
-template <>
-struct algo_level3_blocked_mb_impl<Kokkos::Experimental::HIPSpace> {
-  static constexpr int value = 2;
-};
-#endif
-#if defined(KOKKOS_ENABLE_SYCL)
-template <>
-struct algo_level3_blocked_mb_impl<Kokkos::Experimental::SYCLDeviceUSMSpace> {
-  static constexpr int value = 2;
-};
-#endif
-
-template <class>
-struct algo_level2_blocked_mb_impl;
-template <>
-struct algo_level2_blocked_mb_impl<Kokkos::HostSpace> {
-  static constexpr int value = 4;
-};
-#if defined(KOKKOS_ENABLE_CUDA)
-template <>
-struct algo_level2_blocked_mb_impl<Kokkos::CudaSpace> {
-  static constexpr int value = 1;
-};
-#endif
-#if defined(KOKKOS_ENABLE_HIP)
-template <>
-struct algo_level2_blocked_mb_impl<Kokkos::Experimental::HIPSpace> {
-  static constexpr int value = 1;
-};
-#endif
-#if defined(KOKKOS_ENABLE_SYCL)
-template <>
-struct algo_level2_blocked_mb_impl<Kokkos::Experimental::SYCLDeviceUSMSpace> {
-  static constexpr int value = 1;
-};
-#endif
-
-#endif
-
 using KokkosBlas::Algo;
 
 struct Util {
